add new exceptions and rework them all
